language_demos/requirements.txt
astroid==2.5.1
autopep8==1.5.6
colorama==0.4.4
isort==5.8.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
pycodestyle==2.7.0
pylint==2.7.2
PyYAML==5.4.1
toml==0.10.2
wrapt==1.12.1
